Update Layout Of The Cards
Update the layout of the cards:
- instead of multiple buttons, use one primary button and a dropdown
  menu for the other buttons
- make text of the primary button responsive so on devices with smaller
  displays, a shorter text is used
- use fewer cards per row on iPad-sized devices
- use a badge on the upper right corner of each card to show the
  ReadStatus (wether the volume/issue has been read before)

Already includes javascript and css for the reading mode, as these are
part of shared files between the existing views and the upcoming reading
view.
Also already includes the necessary changes to the index.php and
menu.html files as the VolumeIssuesView contains links to the upcoming
reading mode.

Signed-off-by: ahahn94 <ahahn94@outlook.com>